requirements.txt
aioredis==2.0.0a1
anyio==3.6.1
async-timeout==4.0.2
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Deprecated==1.2.13
fastapi==0.79.0
fastapi-cache2==0.1.8
fastapi-utils==0.2.1
greenlet==1.1.2
h11==0.13.0
hiredis==2.0.0
httptools==0.4.0
idna==3.3
packaging==21.3
pendulum==2.1.2
pydantic==1.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytzdata==2020.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
typing_extensions==4.3.0
urllib3==1.26.10
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.15.0
websockets==10.3
wrapt==1.14.1
